call depositreceipt directly
This is because in the next step, __call__ on the adapter will start
looking at the request method to decide what to do. The deposit receipt
should only be served if it is a GET. Calling the adapter to get a
deposit receipt after doing an upload will call it with a POST request
and it will do the wrong thing. This will do the right thing.