more flexible errordocument/sworderror integration
Errordocument no longer assumes there is a traceback. Instead it takes
its values from verbose and summary. These are fed with proper values
stored on the SwordException classes, or in the case of an unknown
exception, with a traceback. This change allows us to in future simply
raise the required exception, with a verbose description as argument.